+++ R/RcppExports.R
@@ -132,16 +132,16 @@ hwd <- function(ped, reference = NULL) {
 #' @template Fi-arg
 #' @param name_O A string naming the new output column for total opportunity of purging (defaults to "O") 
 #' @param name_Oe A string naming the new output column for the expressed opportunity of purging (defaults to "Oe")
+#' @param sufix A string naming the sufix for non-corrected O and Oe measures
 #' @param compute_O Enable computation of total opportunity of purging (false by default)
-#' @param complex Enable correction for complex pedigrees.
 #' @return The input dataframe, plus two additional column named "O" and "Oe", containing total and expressed opportunity of purging measures.
 #' @encoding UTF-8
 #' @references
 #' \itemize{
 #'   \item{Gulisija D, Crow JF. 2007. Inferring purging from pedigree data. Evolution 61(5): 1043-1051.}
 #' }
-op <- function(ped, pi, Fi, name_O, name_Oe, compute_O = FALSE, complex = TRUE) {
-    .Call(`_purgeR_op`, ped, pi, Fi, name_O, name_Oe, compute_O, complex)
+op <- function(ped, pi, Fi, name_O, name_Oe, sufix, compute_O = FALSE) {
+    .Call(`_purgeR_op`, ped, pi, Fi, name_O, name_Oe, sufix, compute_O)
 }
 
 #' Partial inbreeding coefficient (core function)
@@ -151,13 +151,16 @@ op <- function(ped, pi, Fi, name_O, name_Oe, compute_O = FALSE, complex = TRUE) 
 #' homozygous for alleles derived from ancestor j
 #' 
 #' @name Fij_core_i_cpp
-#' @template ped-arg
+#' @param dam Vector of dam ids.
+#' @param sire Vector of sire ids.
 #' @param anc_idx Index of ancestors.
 #' @param Fi Vector of inbreeding coefficients.
 #' @param mapa Map of ancestors
+#' @param genedrop Enable genedrop simulation
+#' @template seed-arg
 #' @return A matrix of partial inbreeding coefficients. Fi(j) values can thus be read from row i and column j.
-Fij_core_i_cpp <- function(ped, anc_idx, mapa, Fi) {
-    .Call(`_purgeR_Fij_core_i_cpp`, ped, anc_idx, mapa, Fi)
+Fij_core_i_cpp <- function(dam, sire, anc_idx, mapa, Fi, genedrop = 0L, seed = NULL) {
+    .Call(`_purgeR_Fij_core_i_cpp`, dam, sire, anc_idx, mapa, Fi, genedrop, seed)
 }
 
 #' Search and individuals' ancestors