+++ R/purgeR.R
@@ -53,7 +53,10 @@
 #' @aliases purgeR-package
 #' @author Eugenio López-Cortegano <elcortegano@@gmail.com> (\href{https://orcid.org/0000-0001-6914-6305}{ORCID})
 #' @encoding UTF-8
-#' @references López-Cortegano E. 2021. purgeR: Inbreeding and purging in pedigreed populations. *Pending revision*.
+#' @references
+#' \itemize{
+#'   \item{López-Cortegano E. 2021. purgeR: Inbreeding and purging in pedigreed populations. \emph{Bioinformatics}, https://doi.org/10.1093/bioinformatics/btab599.}
+#' }
 #' @seealso
 #' Source code is available via the GitLab repository at \url{https://gitlab.com/elcortegano/purgeR/}. Users are encouraged to report bugs, request features, and contribute code to this project.
 #'